+++ BubbleTweaks/Utilities/AssetLoader.cs
@@ -4,17 +4,17 @@ using UnityEngine;
 
 namespace BubbleTweaks.Utilities {
     class AssetLoader {
-        public static Sprite LoadInternal(string folder, string file) {
-            return Image2Sprite.Create($"{ModSettings.ModEntry.Path}Assets{Path.DirectorySeparatorChar}{folder}{Path.DirectorySeparatorChar}{file}");
+        public static Sprite LoadInternal(string folder, string file, Vector2Int size) {
+            return Image2Sprite.Create($"{ModSettings.ModEntry.Path}Assets{Path.DirectorySeparatorChar}{folder}{Path.DirectorySeparatorChar}{file}", size);
         }
         // Loosely based on https://forum.unity.com/threads/generating-sprites-dynamically-from-png-or-jpeg-files-in-c.343735/
         public static class Image2Sprite {
             public static string icons_folder = "";
-            public static Sprite Create(string filePath) {
+            public static Sprite Create(string filePath, Vector2Int size) {
                 var bytes = File.ReadAllBytes(icons_folder + filePath);
-                var texture = new Texture2D(64, 64, TextureFormat.DXT5, false);
+                var texture = new Texture2D(size.x, size.y, TextureFormat.DXT5, false);
                 _ = texture.LoadImage(bytes);
-                return Sprite.Create(texture, new Rect(0, 0, 64, 64), new Vector2(0, 0));
+                return Sprite.Create(texture, new Rect(0, 0, size.x, size.y), new Vector2(0, 0));
             }
         }
     }