+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 0.2.0
+
+* Added "disband army" button
+
 ## Version 0.1.0
 
-* initial
+* Added speed controls for combat, global map, and tactical battles